user模块登陆注册 -> 迁移 -> auth模块独立登陆注册
