+++ package.json
@@ -25,18 +25,24 @@
   "dependencies": {
     "@nestjs/common": "^9.0.0",
     "@nestjs/core": "^9.0.0",
+    "@nestjs/jwt": "^10.0.1",
+    "@nestjs/passport": "^9.0.0",
     "@nestjs/platform-express": "^9.0.0",
     "@nestjs/platform-socket.io": "^9.2.1",
     "@nestjs/swagger": "^6.1.4",
     "@nestjs/typeorm": "^9.0.1",
     "@nestjs/websockets": "^9.2.1",
     "@prisma/client": "^4.8.0",
+    "@types/passport-local": "^1.0.34",
     "@types/webpack-env": "^1.18.0",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.0",
     "fork-ts-checker-webpack-plugin": "^7.2.14",
     "minecraft-server-util": "^5.3.1",
     "mysql2": "^2.3.3",
+    "passport": "^0.6.0",
+    "passport-jwt": "^4.0.1",
+    "passport-local": "^1.0.0",
     "prisma": "^4.8.0",
     "reflect-metadata": "^0.1.13",
     "run-script-webpack-plugin": "^0.1.1",
@@ -52,6 +58,7 @@
     "@types/express": "^4.17.13",
     "@types/jest": "29.2.4",
     "@types/node": "18.11.18",
+    "@types/passport-jwt": "^3.0.8",
     "@types/supertest": "^2.0.11",
     "@typescript-eslint/eslint-plugin": "^5.0.0",
     "@typescript-eslint/parser": "^5.0.0",
