+++ src/modules/auth/auth.controller.ts
@@ -1,0 +1,19 @@
+import { Controller, Get, Post, Body, Query } from '@nestjs/common';
+import { CreateUserDto } from './dto/create-user.dto';
+import { LoginUserDto } from './dto/login-user.dto';
+import { AuthService } from './auth.service';
+import { QueryUserDto } from './dto/query-user.dto';
+@Controller('auth')
+export class AuthController {
+  constructor(private readonly userService: AuthService) {}
+
+  @Post('signup')
+  async signupUser(@Body() userData: CreateUserDto) {
+    return this.userService.createUser(userData);
+  }
+
+  @Post('login')
+  async loginUser(@Body() userData: LoginUserDto) {
+    return this.userService.loginUser(userData);
+  }
+}
