+++ src/modules/auth/dto/query-user.dto.ts
@@ -1,0 +1,5 @@
+import { IsString } from 'class-validator';
+export class QueryUserDto {
+  @IsString()
+  id: string;
+}
